+++ Vet/src/main/java/ar/com/veterinaria/app/exception/BreedNotFoundException.java
@@ -8,7 +8,7 @@ public class BreedNotFoundException extends RuntimeException {
 
 	private static final long serialVersionUID = 1093306691287711077L;
 
-	public BreedNotFoundException(long id) {
+	public BreedNotFoundException(Integer id) {
 		super("Could not find Breed with id " + id);
 	}
